Extract out non-authorization specific constants and functions
