+++ src/tests/test_authentication.py
@@ -9,7 +9,7 @@ from starlette import status
 
 
 from authentication import get_user_or_raise, keycloak_openid
-from tests.authorization.test_authorization import logged_in_user, ALICE
+from tests.testutils.database import ALICE, logged_in_user
 from tests.testutils.mock_keycloak import MockedKeycloak, TestUserType
